.github/actions/validate-organizations/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "organization-validator"
version = "0.1.0"
description = "Validates organization YAML files for conflicts and schema compliance"
requires-python = ">=3.11"
dependencies = [
    "PyYAML==6.0.2",
    "click==8.1.7",
    "requests==2.32.3",
    "pytest==8.3.3",
    "pytest-mock==3.14.0",
    "responses==0.25.3",
    "ruff>=0.12.9",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short"]
